% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customProbFunctions.R
\name{autoSolution}
\alias{autoSolution}
\alias{fullSolution}
\alias{numericalSolution}
\title{Calculate and itegrate probability functions for shape functions}
\usage{
autoSolution(suffix, shape, tau_0, tau_n, tau, tau_b = NULL, ...,
  MoreArgs = list(), integrateArgs1 = list(),
  integrateArgs2 = list(), forceNumeric = F)

fullSolution(suffix, shape, tau_0, tau_n, tau, tau_b = NULL, ...,
  MoreArgs = list())

numericalSolution(suffix, shape, tau_0, tau_n, tau, tau_b = NULL, ...,
  MoreArgs = list(), integrateArgs1 = list(),
  integrateArgs2 = list(), forceNumeric = T)
}
\arguments{
\item{suffix}{A character vector indicating the value to be returned; either
"PDF", "CCDF", "IntPDF", "IntCCDF", "IntTau.PDF", or "IntTau.CCDF."}

\item{shape}{A character vector of \code{length(shape) == 1} naming the shape
function (e.g., \code{\link{powerLaw}}) to be integrated numerically if
solutions are not available used.}

\item{tau_0, tau_n}{Minimum and maximum residence times of the definite
integration.}

\item{tau}{A vector of residence times for which PDF or CCDF values are
requested, or the lower values of definate intergrals of the finite
integral of the PDF, CCDF, tau*PDF, or tau*CCDF.}

\item{tau_b}{The upper values of definite integrals of the PDF, CCDF,
tau*PDF, or tau*CCDF.  Ignored for calculating PDF or CCDF values.}

\item{...}{Additional values required by the shape function.}

\item{integrateArgs1, integrateArgs2}{When numerical integration is used, a
named list of optional values for \code{\link[stats]{integrate}} function
(e.g., \code{subdivisions, rel.tol}, etc.) can be passed via this argument.
Any values in integrateArgs1 list will be applied to the integration used
within the calculations for the PDF or CCDF.  Values for integrateArgs2
will be used when the PDF or CCDF are integrated.  Thus, integrateArgs1 are
used regardless of the value of \code{suffix}.  integrateArgs2 are used
only when \code{suffix} is "IntPDF", "IntCCDF", "IntTau.PDF", or
"inteTau.CCDF".  When solutions functions are used (rather than numerical
integration), these parameters are ignored.}

\item{forceNumeric}{When set to T, numerical integration will be used by
\code{autoSolution} even if solutions to the PDF, CCDF, IntPDF, IntCCDF,
IntTau.PDF, and IntTau.CCDF are available.}
}
\value{
A vector of values representing either the PDF, CCDF, integration of
  the PDF, integration of the CCDF, integration of \code{tau} * PDF or the
  integration of \code{tau} * CCDF.
}
\description{
Given the name of a shape function and the specific calculation requested,
calculate either the PDF, CCDF, integration of the PDF or CCDF, or the
integration of tau*PDF or tau*CCDF.
}
\details{
The function \code{numericalSolution} uses numerical integration of basic
shape function (e.g., \code{\link{powerLaw}}, \code{\link{exponent}}) to return
values of the probability density function (PDF) of the shape function,
complementary cumulative distribution funtion (CCDF) of the shape function,
finite integrals of the PDF or CCDF, or finite integrals of \code{tau*PDF} or
\code{tau*CCDF}.  Numerical integration to calculate the normalizing constant
for the PDF and CCDF, and for integrating these functions.

The function \code{fullSolution} looks for a function with a name specified
by the concatination of \code{shape} and \code{suffix} and checks to be sure
any such function has the signature \code{function(tau,} \code{tau_0,} \code{tau_n, ...)}
if suffix is "PDF" or "CCDF", or \code{function(tau_a,} \code{tau_b,} \code{tau_0,} \code{tau_n,}
\code{...)} for other values of suffix.  Solutions for the "powerLaw" and "exponent"
shapes are provided as part of this package as follows:  PDF (e.g.
\code{\link{powerLawPDF}}, \code{\link{exponentPDF}}), CCDF (e.g.
\code{\link{powerLawCCDF}}, \code{\link{exponentCCDF}}), integral of the PFD (e.g.
\code{\link{powerLawIntPDF}}, \code{\link{exponentIntPDF}}), integral of the CCDF
(e.g. \code{\link{powerLawIntCCDF}}, \code{\link{exponentIntCCDF}}), integral of
tau*PDF (e.g. \code{\link{powerLawIntTau.PDF}}, \code{\link{exponentIntTau.PDF}}) or
integral of tau*CCDF (e.g. \code{\link{powerLawIntTau.CCDF}},
\code{\link{exponentIntTau.CCDF}}).  These solution functions can also be called
directly, but \code{fullSolution} is provided as a convenience wrapper that
calls the functions using the convention \code{shape} and \code{suffix}.

The user can provide custom shape functions or solutions for other shapes,
however the names of the custom solution functions must follow the convention
of concatinating the shape name with the suffix as the name of the solution
function (e.g., the solution function for the integral of tau*PDF of a shape
called "foo" must be \code{fooIntTau.PDF}).  For more information, see the
documentation of \code{\link{powerLaw}} and \code{\link{powerLawPDF}}

The function \code{autoSolution} looks first for a solution function that
matches the concatination of \code{shape} and \code{suffix} and, if found,
calls \code{fullSolution}. If the solution functions are not found,
\code{autoSolution} calls \code{numericSolution}.
}
