% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customProbFunctions.R
\name{powerLaw}
\alias{powerLaw}
\alias{exponent}
\title{Example shape functions}
\usage{
powerLaw(tau, tau_n, alpha)

exponent(tau, tau_n, sigma)
}
\arguments{
\item{tau}{A numeric vector of residence times.}

\item{tau_n}{A numeric vector (usually but not necessarily lenth == 1)
representing the maximum \code{tau} of the hyporehic zone.}

\item{alpha}{The exponent of the power law; this value is negated within the
\code{powerLaw} function.}

\item{sigma}{The hydrologic exchange rate for the \code{exponent} function;
this value is negated within the \code{exponent} function.}
}
\value{
Shape functions (including custom shape functions you write) must
  recycle values of \code{tau} and \code{tau_n}. Effectively, since the
  \code{tau_n} is typically length == 1, shape functions usually return a
  vector of numeric values equal in length to the \code{tau} vector passed to
  the function. Both \code{powerLaw} and \code{exponent} conform to these
  requirements. The \code{powerLaw} function returns \code{tau^(-alpha) -
  tau_n^(-alpha)}.  The \code{exponent} function returns
  \code{exp(-sigma*tau) - exp(-sigma * tau_n)}.  Implementation of these
  functions can be viewed by typing \code{print(powerLaw)} or
  \code{print(exponent)} into the console.
}
\description{
Transient storage zone statistics are calucated based on an assumed shape of
the hydrologic residence time distribution (RTD) of the hyporheic zone, which
is the probablity density function (PDF) that a water molecule will exit the
hyporheic zone at time \code{tau}.  \code{powerLaw} and \code{exponent} are
examples of functions that describe a basic shape for the RTD.
}
\details{
Given an appropriate "shape function," \code{HydroGeom} can calculate
transient storage zone statistics for any assumed shape of the RTD. Shape
functions must have the signature \code{function(tau, tau_n, ...)} and
typically return zero when tau == tau_n.

You can generate hyporheic TSZ statistics based on a RTD with a shape
corresponding to a custom function that you provide.  If you write a shape
function like \code{foo = function(tau, tau_n, chi, omega)}, where \code{chi}
and \code{omega} are parameters that influence the shape of the curve
described by your function (e.g., analagous to the way \code{alpha} serves as
an exponent that controls the specific shape of a power law curve in the
\code{powerLaw} function), you can then use \code{\link{TSZStats}} with
\code{shape = "foo"} to generate TSZ statistics using a custom shape for the
RTD of the hyporheic zone.  Implemented this way, results of \code{\link{TSZStats}}
are derived using numerical integration (see \code{\link{numericalSolution}}).
Solutions for integration of a custom function can also be provide (see \code{\link{fullSolution}} and
\code{\link{powerLawPDF}}).

Technically speaking, \code{tau_n} serves as the upper limit of \code{tau}
used to calculate the normalizing constant of the PDF, which is defined as
the inverse of the definite integral of the shape function from some lower
limit (\code{tau_0}) to tau_n.  Pragmatically, \code{tau_n} represents the
maximum residence time of interest, which should approximate the maximum
subsurface residence time of hyporheic water.  Water that stays in the
subsurface longer than \code{tau_n} is assumed to enter the true groundwater
system.

All paramerters for these fuctions correspond to the defintions found in
Poole et al. (In Press) "The Hydrolgic Geometry of Hyporheic Zones in
Expansive Coarse-Grained Alluvial Aquifers."
}
\seealso{
\code{\link{checkShapeFunction}}
}
