% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSZStats.R
\name{TSZStats}
\alias{TSZStats}
\alias{hyporheicBins}
\title{Generate transient storage statistics}
\usage{
TSZStats(TSZs, tau_0, tau_n, storage = NULL, q = NULL, ..., factor = NULL,
  shape = "powerLaw", MoreArgs = list(), integrateArgs1 = list(),
  integrateArgs2 = list(), forceNumeric = F, optimizeInterval = NULL,
  optimizeTol = .Machine$double.eps^0.25)

hyporheicBins(nbins, factor, minRT, maxRT, porosity, hyporheicSize,
  hyporheicExchange = NULL, b = NULL, integrateArgs = list())
}
\arguments{
\item{TSZs}{Either the number of transient storage zones requested, or a
vector of residence times that serve as the breakpoints among transient
stroage zones.}

\item{tau_0}{The smallest residence time to be considered part of the
hyporheic zone.}

\item{tau_n}{The largest residence time to be considered part of the
hyporehic zone.}

\item{storage}{Amount of water stored in the hyporheic zone.}

\item{q}{Gross hyporheic exchange per time unit.  Exchange can be unit can be
distance, area, or volume per unit time, but must be the same as used units
used to describe storage.  Time unit must be the same as units used to
describe tau_0 and tau_n.}

\item{factor}{Controls the distribution of residence time for transient
storage zones when \code{TSZs} is set equal to the number of transient
storage zones. Value must be either 0 or >1. See details.  \code{factor} is
ignored when TSZs is a vector of the residence time breakpoints between
transient storage zones.}

\item{shape}{A character string naming the shape file (e.g.,
\code{\link{powerLaw}}) to be used in the calculation of transient storage
statistics.}

\item{MoreArgs}{A named list of additional parameter values to be passed to
the shape function -- only valid when numerical integration is used;
ignored when solutions to the shape function are used.  When passed this
way, the values will not be vectorized with tau, tau_0, tau_n, tau_b, or
arguments in '...'.}

\item{integrateArgs1, integrateArgs2}{When numerical integration is used,
these parameters are a named list of optional values for the
\code{\link[stats]{integrate}} function.  For PDF and CCDF calculations,
integrateArgs1 is used and integrateArgs2 is ignored.  For the integration
of the PDF, CCDF, tau*PDF or tau*CCDF, integrateArgs1 will apply to the
internal call to the PDF or CCDF functions and integrateArgs2 will apply to
the subsequent integration of those functions.  In practise, use of these
parameters is seldom required.  These parameters are useful only if and
unexpected error message is returned by the \code{\link{integrate}}
function.}

\item{....}{Additional arguments required by shape function (e.g.,
\code{alpha} for \code{\link{powerLaw}}).  Must be named.  First of these
values can be passed as NULL and will be estimated if both \code{storage}
and \code{q} are provided.  See details.}
}
\value{
A TSZ object, which is a data.frame with attributes that record the
  arguments (\code{TSZs, tau_0, tau_n, storage, q, factor, and forceNumeric},
  along with any shape specific parameters (e.g., \code{alpha} for
  \code{\link{powerLaw}})) used to calculate the values in the data.frame.
  Attributes also include the descrepancy between the request and calculated
  q and storage. Descrepancy values should be
}
\description{
The exchange of water among and size of multiple transient storage zones
within the hyporehic zone are calculated according to the method describe in
Poole et al. (In press).  Calculations are made either by utilizing solutions
to various integrations of a shape function (e.g., \code{\link{powerLawPDF}}
and associated fuctions) or via numerical integration of a shape fuction
(e.g., \code{\link{powerLaw}}).
}
\details{
When \code{factor == 0}, transient storage zone breakpoints will be
determined such that each transient storage zone has and equal amount of
water storage.  When factor is 1 or greater, \code{factor} is used as a
mutiplier to determine the residence times associated with each TSZ such that
the range of residence times contained in any TSZ is \code{<factor>} times
larger than the range of residence times contained in the previous TSZ.  So
if \code{factor == 2}, the range of residence times in each TSZ will be 2x
the range of the prior TSZ.

If passed as NULL, the value for first parameter after \code{tau} and
\code{tau_n} in the formals of the function named by \code{shape} can be
estimated by \code{TSZStats}, however non-NULL values for both \code{storage}
and \code{q} must be provided.  For instance, for the function
\code{\link{powerLaw}}, a value for \code{alpha} will be estimated by
TSZStats if \code{alpha = NULL} is included in the call to \code{TSZStats}
along with non-NULL values for \code{storege} and \code{q}.

\code{HyporheicBins} has been depricated and is now a wrapper for
\code{TSZStats}.  Use \code{TSZStats} instead.
}

